requirements.txt
streamlit==1.28.0
opencv-python==4.8.1.78
mediapipe==1.13.0
numpy==1.25.2
Pillow==10.0.0
scikit-learn==1.3.0
joblib==1.3.2